fix: eliminate all compiler warnings
- Fix -Wconversion warnings by adding explicit static_cast<double> for division operations
- Remove unused feed_lines function to eliminate -Wunused-function warning
- Fix unused parameter warnings by using /* parameter_name */ syntax
- Fix member initialization order warnings by reordering member variables
- Fix conversion warnings in serial and tcp_client by casting to unsigned int

All compiler warnings have been eliminated while maintaining functionality.

+++ unilink/common/logger.hpp
@@ -105,7 +105,7 @@ struct AsyncLogStats {
 
   double get_drop_rate() const {
     if (total_logs == 0) return 0.0;
-    return static_cast<double>(dropped_logs) / total_logs;
+    return static_cast<double>(dropped_logs) / static_cast<double>(total_logs);
   }
 
   std::chrono::milliseconds get_uptime() const {
